api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
aiosqlite==0.20.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
httpx==0.27.2
python-multipart==0.0.12
websockets==13.1
psutil==6.0.0
pyyaml==6.0.2
